requirements.txt
bleak==0.21.1
certifi==2024.2.2
charset-normalizer==3.3.2
dotmap==1.3.30
idna==3.6
meshtastic==2.3.3
numpy==1.26.4
packaging==24.0
pexpect==4.9.0
prometheus_client==0.20.0
protobuf==5.26.1
ptyprocess==0.7.0
pubsub==0.1.2
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-CoreBluetooth==9.2
pyobjc-framework-libdispatch==9.2
Pypubsub==4.0.3
PyQRCode==1.2.1
pyserial==3.5
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tabulate==0.9.0
timeago==1.0.16
typing_extensions==4.10.0
urllib3==2.2.1
